src/app/components/volume-control/volume-control.component.ts
import { Component, OnInit } from '@angular/core';
import { BasePlaybackService } from '../../services/playback/base-playback.service';

@Component({
    selector: 'app-volume-control',
    templateUrl: './volume-control.component.html',
    styleUrls: ['./volume-control.component.scss'],
})
export class VolumeControlComponent implements OnInit {
    constructor(private playbackService: BasePlaybackService) {}

    public get volume(): number {
        return this.playbackService.volume;
    }

    public set volume(v: number) {
        this.playbackService.volume = v;
    }

    public ngOnInit(): void {}
}
